arossupport.library: Fix some compiler complaints on 64-bit

git-svn-id: https://svn.aros.org/svn/aros/trunk@37159 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/arossupport/rt_dos.c
@@ -141,7 +141,7 @@ static const STRPTR GetFileMode (LONG mode)
     case MODE_READWRITE: return "MODE_READWRITE";
     }
 
-    sprintf (buffer, "<illegal mode %ld>", mode);
+    sprintf (buffer, "<illegal mode %ld>", (long)mode);
 
     return buffer;
 } /* GetFileMode */
